Code cleanup and move some tests
